Refactor: Separate WebSocket and HTTP functionalities into modular components
