+++ server/http_server.go
@@ -10,14 +10,11 @@ import (
 func StartHTTPServer(chatHandler *api.ChatHandler, authHandler *api.AuthHandler) {
     router := gin.Default()
 
-    router.Static("/static", "./frontend")
-
     apiGroup := router.Group("/api")
     {
         apiGroup.POST("/signup", authHandler.SignupGin)
         apiGroup.POST("/login", authHandler.LoginGin)
         apiGroup.GET("/messages", chatHandler.GetMessagesGin)
-        apiGroup.GET("/group/messages?group_id=",chatHandler.GetGroupMessages)
     }
 
     log.Println("HTTP server started on :8080")
