+++ server/websocket/module.go
@@ -1,0 +1,7 @@
+package websocket
+
+import "github.com/Talal52/go-chat/chat/service"
+
+func NewWebSocketModule(chatService *service.ChatService) *WebSocketServer {
+    return NewWebSocketServer(chatService)
+}
